Move enum into separate file
